1.2.0.7

+++ zp-core/zp-extensions/auto_backup.php
@@ -30,6 +30,7 @@ if (OFFSET_PATH == 2) {
 		require_once(dirname(dirname(__FILE__)) . '/admin-functions.php');
 		zp_register_filter('admin_head', 'auto_backup::timer_handler');
 		zp_register_filter('theme_head', 'auto_backup::timer_handler');
+		$_backupMutex = new zpMutex('bK');
 	}
 }
 
@@ -80,27 +81,34 @@ class auto_backup {
 	 * @param string $discard
 	 */
 	static function timer_handler($discard) {
-		$curdir = getcwd();
-		$folder = SERVERPATH . "/" . BACKUPFOLDER;
-		if (!is_dir($folder)) {
-			mkdir($folder, FOLDER_MOD);
+		global $_backupMutex;
+		$_backupMutex->lock();
+		if ((getOption('last_backup_run') + getOption('backup_interval') * 86400) < time()) {
+			//	maybe a race condition? Only need one execution
+			$curdir = getcwd();
+			$folder = SERVERPATH . "/" . BACKUPFOLDER;
+			if (!is_dir($folder)) {
+				mkdir($folder, FOLDER_MOD);
+			}
+			chdir($folder);
+			$filelist = safe_glob('*' . '.zdb');
+			$list = array();
+			foreach ($filelist as $file) {
+				$list[$file] = filemtime($file);
+			}
+			chdir($curdir);
+			asort($list);
+			$list = array_flip($list);
+			$keep = getOption('backups_to_keep');
+			while (!empty($list) && count($list) >= $keep) {
+				$file = array_shift($list);
+				@chmod(SERVERPATH . "/" . BACKUPFOLDER . '/' . $file, 0777);
+				unlink(SERVERPATH . "/" . BACKUPFOLDER . '/' . $file);
+			}
+			cron_starter(SERVERPATH . '/' . ZENFOLDER . '/' . UTILITIES_FOLDER . '/backup_restore.php', array('backup' => 1, 'autobackup' => 1, 'compress' => sprintf('%u', getOption('backup_compression')), 'XSRFTag' => 'backup'), 3);
+			setOption('last_backup_run', time());
 		}
-		chdir($folder);
-		$filelist = safe_glob('*' . '.zdb');
-		$list = array();
-		foreach ($filelist as $file) {
-			$list[$file] = filemtime($file);
-		}
-		chdir($curdir);
-		asort($list);
-		$list = array_flip($list);
-		$keep = getOption('backups_to_keep');
-		while (!empty($list) && count($list) >= $keep) {
-			$file = array_shift($list);
-			@chmod(SERVERPATH . "/" . BACKUPFOLDER . '/' . $file, 0777);
-			unlink(SERVERPATH . "/" . BACKUPFOLDER . '/' . $file);
-		}
-		cron_starter(SERVERPATH . '/' . ZENFOLDER . '/' . UTILITIES_FOLDER . '/backup_restore.php', array('backup' => 1, 'autobackup' => 1, 'compress' => sprintf('%u', getOption('backup_compression')), 'XSRFTag' => 'backup'), 3);
+		$_backupMutex->unlock();
 		return $discard;
 	}
